Rename utility modules and update import paths for consistency
